% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{koans_df}
\alias{koans_df}
\title{Ordered list of the lessons in advRkoans along with summary statistics}
\format{A data frame with 2 rows and 3 variables:
\describe{
  \item{lesson_file}{relative location of the lesson file}
  \item{number_koans}{number of koans in the lesson}
  \item{cumulative_koans}{number of cumulative koans through that lesson}
  \item{test_file}{relative location of the test file}
}}
\usage{
koans_df
}
\description{
Ordered list of the lessons in advRkoans along with summary statistics
}
\keyword{datasets}
