% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contemplate_koans.R
\name{walk_the_path}
\alias{walk_the_path}
\title{Walk the path to R enlightenment}
\usage{
walk_the_path(df = koans_df, koans_passed = 0, lessons_passed = 0)
}
\description{
\code{walk_the_path} runs tests within the lessons and provides feedback
}
